refactor: replace Navbar component with PublicNavbar and update imports across multiple pages
