+++ src/app/(public)/privacy/page.tsx
@@ -1,4 +1,4 @@
-import Navbar from "@/components/Navbar";
+import Navbar from "@/components/navbar/PublicNavbar";
 import Footer from "@/components/Footer";
 
 export default function Privacy() {
